add syslog and A1/A2 trigger
Add optional logging to a SysLog server.

Right now all HTTP endpoints are logged as well as
the pin actions that are started from the endpoints.

Use A1 and A2 pins as inputs (pulled up) and check
for a signal pulse on them. In response, pulse the 1
or 2 pins, respectively.

This is used to allow another device to pulse the output
pins while also noting this event in the log.